fix duplicated websocket tests after merge...
